pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taskflow-api"
version = "0.1.0"
description = "Taskflow API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.2.1",
  "langgraph>=0.2.34",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "black==24.8.0",
  "ruff>=0.6.0,<1.0",
  "bandit>=1.7.9,<2.0",
  "pip-audit>=2.7.3,<3.0"
]
llm-openai = [
  "langchain-openai>=0.2.0"
]
llm-anthropic = [
  "langchain-anthropic>=0.2.0"
]
llm-providers = [
  "langchain-openai>=0.2.0",
  "langchain-anthropic>=0.2.0"
]

[tool.setuptools]
py-modules = ["main"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
